Frontend Inventories Split with Player and Shelf (Needs Fix)
Modified the shelf_ui to handle and grab the player inventory and the shelves. However slots within their respective inventories are persistant, but across is not.

FULL NOTE:
so switching items within their respective slots works fine, but once you switch items between slots it gets weird. Sometimes the switching between inventories works and sometimes it doesnt (just play around with switching back and forth and you'll see). its kind of like the inventory is not in sync with the ui. the shelf_inv_ui is like yours but is modified to handle two inventories that are treated as one big slot array (0-4 is player, 5 - 16 is shelf).

Also I never set the player inventory so its not reflected on the player inventory, im guessing cause slots themselves don't call like player.collect but not 100% sure.